LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/popovertarget-expected.txt

Toggle the popover  Toggle the popover  Toggle the popover

FAIL Shadow root reference target works with popovertarget attribute. assert_equals: host should be toggled if reference target is null expected 1 but got 0
FAIL Shadow root reference target works with popovertarget attribute via options. assert_equals: host should be toggled if reference target is null expected 1 but got 0
FAIL Shadow root reference target works with .popoverTargetElement property. assert_equals: host should be toggled if reference target is null expected 1 but got 0
